Multiplication result converted to larger type (https://github.com/libxsmm/libxsmm/security/code-scanning).

+++ src/generator_gemm_common_aarch64.c
@@ -272,7 +272,7 @@ void libxsmm_generator_gemm_aarch64_setup_k_strides( libxsmm_generated_code*    
 
   /* preload offset of A */
   libxsmm_aarch64_instruction_alu_set_imm64( io_generated_code, i_gp_reg_mapping->gp_reg_help_0,
-                                           (unsigned long long)((unsigned long long)(i_xgemm_desc->lda - i_m_blocking) * i_micro_kernel_config->datatype_size_in) );
+                                           ((unsigned long long)i_xgemm_desc->lda - i_m_blocking) * i_micro_kernel_config->datatype_size_in );
 
   /* load b offsets */
   if ( io_generated_code->arch != LIBXSMM_AARCH64_A64FX ) {
